load all existing player when landing

+++ src/main/java/com/guxuede/gm/net/userdata/UserDto.java
@@ -4,7 +4,7 @@ public class UserDto {
     private String userName;
     private int id;
     private String character;
-    private int x,y;
+    private float x,y;
     private int direction;
 
     public String getUserName() {
@@ -31,19 +31,19 @@ public class UserDto {
         this.character = character;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
